refactor: move back off into separate module
